Added 3 files that allow us to fetch student data
the index calls both the curl2.php and getStudData.js to scrape student
data from http://www.ischool.berkeley.edu/people/students. The data is
cleaned with regex then formatted in an object as: {name: [name, year,
interests]}
Note that this only works if you place the file on the ischool drive. I
have it here for the time being:
http://people.ischool.berkeley.edu/~nkhalawi/projects/iolab_proj4/